+++ core/src/test/java/com/datastrato/gravitino/catalog/TestBaseSchema.java
@@ -2,7 +2,7 @@
  * Copyright 2023 Datastrato Pvt Ltd.
  * This software is licensed under the Apache License version 2.
  */
-package com.datastrato.gravitino.catalog.rel;
+package com.datastrato.gravitino.catalog;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
