+++ core/src/test/java/com/datastrato/gravitino/metalake/TestMetalakeManager.java
@@ -2,7 +2,7 @@
  * Copyright 2023 Datastrato Pvt Ltd.
  * This software is licensed under the Apache License version 2.
  */
-package com.datastrato.gravitino.meta;
+package com.datastrato.gravitino.metalake;
 
 import com.datastrato.gravitino.Config;
 import com.datastrato.gravitino.EntityStore;
@@ -12,6 +12,7 @@ import com.datastrato.gravitino.StringIdentifier;
 import com.datastrato.gravitino.TestEntityStore;
 import com.datastrato.gravitino.exceptions.MetalakeAlreadyExistsException;
 import com.datastrato.gravitino.exceptions.NoSuchMetalakeException;
+import com.datastrato.gravitino.meta.BaseMetalake;
 import com.datastrato.gravitino.storage.RandomIdGenerator;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Sets;
