+++ server/src/test/java/com/datastrato/gravitino/server/web/rest/TestMetalakeOperations.java
@@ -28,8 +28,8 @@ import com.datastrato.gravitino.exceptions.NoSuchMetalakeException;
 import com.datastrato.gravitino.lock.LockManager;
 import com.datastrato.gravitino.meta.AuditInfo;
 import com.datastrato.gravitino.meta.BaseMetalake;
-import com.datastrato.gravitino.meta.MetalakeManager;
 import com.datastrato.gravitino.meta.SchemaVersion;
+import com.datastrato.gravitino.metalake.MetalakeManager;
 import com.datastrato.gravitino.rest.RESTUtils;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
